removed non-skorch AMP calculator and trainer and corresponding tests. Skorch remains default trainer.
