+++ amptorch/tests/val_test.py
@@ -3,8 +3,6 @@ from torch import optim
 from skorch import NeuralNetRegressor
 from skorch.dataset import CVSplit
 from skorch.callbacks import Checkpoint, EpochScoring
-from amptorch.core import AMPTorch
-from amptorch import AMP
 from amptorch.gaussian import SNN_Gaussian
 from amptorch.skorch_model import AMP as AMP_skorch
 from amptorch.skorch_model.utils import target_extractor, energy_score, forces_score
@@ -218,197 +216,3 @@ def test_energy_only_skorch_val():
     assert round(energy_rmse, 5) == round(
         last_energy_score, 5
     ), "Validation energy errors incorrect!"
-
-def test_val():
-    distances = np.linspace(2, 5, 10)
-    label = "example"
-    images = []
-    energies = []
-    forces = []
-    for l in distances:
-        image = Atoms(
-            "CuCO",
-            [
-                (-l * np.sin(0.65), l * np.cos(0.65), 0),
-                (0, 0, 0),
-                (l * np.sin(0.65), l * np.cos(0.65), 0),
-            ],
-        )
-        image.set_cell([10, 10, 10])
-        image.wrap(pbc=True)
-        image.set_calculator(EMT())
-        images.append(image)
-        energies.append(image.get_potential_energy(apply_constraint=False))
-        forces.append(image.get_forces(apply_constraint=False))
-
-    energies = np.array(energies)
-    forces = np.concatenate(np.array(forces))
-    Gs = {}
-    Gs["G2_etas"] = np.logspace(np.log10(0.05), np.log10(5.0), num=2)
-    Gs["G2_rs_s"] = [0] * 2
-    Gs["G4_etas"] = [0.005]
-    Gs["G4_zetas"] = [1.0]
-    Gs["G4_gammas"] = [+1.0, -1]
-    Gs["cutoff"] = 6.5
-
-    torch.set_num_threads(1)
-    calc = AMP(
-        model=AMPTorch(
-            images,
-            descriptor=SNN_Gaussian,
-            Gs=Gs,
-            val_frac=0.2,
-            force_coefficient=0.3,
-            label=label,
-            save_logs=True,
-        )
-    )
-    calc.model.device = "cpu"
-    calc.model.structure = [2, 2]
-    calc.model.convergence = {
-        "energy": 0.3,
-        "force": 0.3,
-        "early_stop": False,
-        "epochs": 20,
-    }
-    calc.model.loader_params = {"batch_size": None, "shuffle": False, "num_workers": 0}
-    calc.model.criterion = CustomMSELoss
-    calc.model.optimizer = optim.LBFGS
-    calc.model.lr = 1e-2
-    calc.model.fine_tune = None
-
-    calc.train(overwrite=True)
-    loaders = calc.model.atoms_dataloader
-    val_loader = loaders["val"]
-    val_idx = val_loader.sampler.indices
-    validation_images = [images[idx] for idx in val_idx]
-    val_energies = np.array(
-        [
-            image.get_potential_energy(apply_constraint=False)
-            for image in validation_images
-        ]
-    )
-    val_forces = np.concatenate(
-        np.array(
-            [image.get_forces(apply_constraint=False) for image in validation_images]
-        )
-    )
-
-    f_read = open("./results/logs/example.txt", "r")
-    lines = f_read.readlines()
-    energy_line = lines[-3]
-    forces_line = lines[-2]
-    reported_energy_loss = float(energy_line.split()[-1])
-    reported_forces_loss = float(forces_line.split()[-1])
-    f_read.close()
-
-    num_of_atoms = 3
-    calculated_energies = np.array(
-        [calc.get_potential_energy(image) for image in validation_images]
-    )
-    energy_rmse = np.sqrt(
-        (((calculated_energies - val_energies) / num_of_atoms) ** 2).sum()
-        / len(validation_images)
-    )
-    assert round(energy_rmse, 4) == round(
-        reported_energy_loss, 4
-    ), "Energy errors are incorrect!"
-
-    calculated_forces = np.concatenate(
-        np.array([calc.get_forces(image) for image in validation_images])
-    )
-    force_rmse = np.sqrt(
-        (((calculated_forces - val_forces)) ** 2).sum()
-        / (3 * num_of_atoms * len(validation_images))
-    )
-    assert round(force_rmse, 4) == round(
-        reported_forces_loss, 4
-    ), "Force errors are incorrect!"
-
-def test_energy_only_val():
-    distances = np.linspace(2, 5, 10)
-    label = "example"
-    images = []
-    energies = []
-    forces = []
-    for l in distances:
-        image = Atoms(
-            "CuCO",
-            [
-                (-l * np.sin(0.65), l * np.cos(0.65), 0),
-                (0, 0, 0),
-                (l * np.sin(0.65), l * np.cos(0.65), 0),
-            ],
-        )
-        image.set_cell([10, 10, 10])
-        image.wrap(pbc=True)
-        image.set_calculator(EMT())
-        images.append(image)
-        energies.append(image.get_potential_energy(apply_constraint=False))
-        forces.append(image.get_forces(apply_constraint=False))
-
-    energies = np.array(energies)
-    forces = np.concatenate(np.array(forces))
-    Gs = {}
-    Gs["G2_etas"] = np.logspace(np.log10(0.05), np.log10(5.0), num=2)
-    Gs["G2_rs_s"] = [0] * 2
-    Gs["G4_etas"] = [0.005]
-    Gs["G4_zetas"] = [1.0]
-    Gs["G4_gammas"] = [+1.0, -1]
-    Gs["cutoff"] = 6.5
-
-    torch.set_num_threads(1)
-    calc = AMP(
-        model=AMPTorch(
-            images,
-            descriptor=SNN_Gaussian,
-            Gs=Gs,
-            val_frac=0.2,
-            force_coefficient=0,
-            label=label,
-            save_logs=True,
-        )
-    )
-    calc.model.device = "cpu"
-    calc.model.structure = [2, 2]
-    calc.model.convergence = {
-        "energy": 0.1,
-        "force": 0.1,
-        "early_stop": False,
-        "epochs": 20,
-    }
-    calc.model.loader_params = {"batch_size": None, "shuffle": False, "num_workers": 0}
-    calc.model.criterion = CustomMSELoss
-    calc.model.optimizer = optim.LBFGS
-    calc.model.lr = 1e-2
-    calc.model.fine_tune = None
-
-    calc.train(overwrite=True)
-    loaders = calc.model.atoms_dataloader
-    val_loader = loaders["val"]
-    val_idx = val_loader.sampler.indices
-    validation_images = [images[idx] for idx in val_idx]
-    val_energies = np.array(
-        [
-            image.get_potential_energy(apply_constraint=False)
-            for image in validation_images
-        ]
-    )
-
-    f_read = open("./results/logs/example.txt", "r")
-    lines = f_read.readlines()
-    energy_line = lines[-2]
-    reported_energy_loss = float(energy_line.split()[-1])
-    f_read.close()
-
-    num_of_atoms = 3
-    calculated_energies = np.array(
-        [calc.get_potential_energy(image) for image in validation_images]
-    )
-    energy_rmse = np.sqrt(
-        (((calculated_energies - val_energies) / num_of_atoms) ** 2).sum()
-        / len(validation_images)
-    )
-    assert round(energy_rmse, 4) == round(
-        reported_energy_loss, 4
-    ), "Energy errors are incorrect!"
